+++ float/src/repr.rs
@@ -2,7 +2,6 @@ use crate::utils::get_precision;
 use dashu_int::{IBig, Sign, Word};
 
 // TODO: add standalone basic arith methods (add, sub, mul, div) for FloatRepr, such that it returns a Approximation struct
-// TODO: determine whether supporting Infs? Inf can be stored as zero mantissa with nonzero exponent
 
 /// An arbitrary precision floating number represented as `mantissa * radix^exponent`, with a precision
 /// such that `|mantissa| < radix^precision`. The mantissa is also called significant. `Radix` should be
@@ -13,8 +12,20 @@ use dashu_int::{IBig, Sign, Word};
 /// rounding mode. Note that the rounding is only for operations, it's not "associated" with the value.
 /// For example, for correct subtraction, the two operands should have reverse rounding direction.
 ///
+/// # Generic Parameters
 /// The const generic parameters will be abbreviated as Radix -> X, Rounding -> R.
 /// Radix should be in range \[2, isize::MAX\], and Rounding value has to be one of [RoundingMode]
+/// 
+/// # Infinity
+/// 
+/// This struct supports representation the infinity, but the infinity is only supposed to be used as sentinels.
+/// That is, only equality test and comparison are implemented for infinity. Any other operations performed
+/// with infinity will lead to panic.
+/// 
+/// The infinities are represented as:
+/// * (Positive) infinity: mantissa = 0, exponent > 0
+/// * Negative infinity: mantissa = 0, exponenet < 0
+/// 
 #[allow(non_upper_case_globals)]
 #[derive(Clone, Debug)]
 pub struct FloatRepr<const Radix: usize, const Rounding: u8> {
@@ -56,10 +67,15 @@ impl<const X: usize, const R: u8> FloatRepr<X, R> {
         }
     }
 
-    pub const fn from_parts_const(sign: Sign, man_lo: Word, man_hi: Word, mut exponent: isize) -> Self {
+    pub const fn from_parts_const(
+        sign: Sign,
+        man_lo: Word,
+        man_hi: Word,
+        mut exponent: isize,
+    ) -> Self {
         let mut mantissa = (man_hi as u128) << Word::BITS | (man_lo as u128);
         if mantissa == 0 {
-            return Self::zero()
+            return Self::ZERO;
         }
 
         let mut precision = 0;
@@ -88,7 +104,11 @@ impl<const X: usize, const R: u8> FloatRepr<X, R> {
 
         let low = (mantissa & Word::MAX as u128) as Word;
         let high = (mantissa >> Word::BITS) as Word;
-        Self { mantissa: IBig::from_parts_const(sign, low, high), exponent, precision }
+        Self {
+            mantissa: IBig::from_parts_const(sign, low, high),
+            exponent,
+            precision,
+        }
     }
 
     /// Convert raw parts into a float number, with given precision.
@@ -103,19 +123,51 @@ impl<const X: usize, const R: u8> FloatRepr<X, R> {
         (self.mantissa, self.exponent)
     }
 
-    #[inline]
-    pub const fn zero() -> Self {
-        Self { mantissa: IBig::zero(), exponent: 0, precision: 1 }
+    /// [FloatRepr] with value 0
+    pub const ZERO: Self = Self {
+        mantissa: IBig::ZERO,
+        exponent: 0,
+        precision: 1,
+    };
+    /// [FloatRepr] with value 1
+    pub const ONE: Self = Self {
+        mantissa: IBig::ONE,
+        exponent: 0,
+        precision: 1,
+    };
+    /// [FloatRepr] with value -1
+    pub const NEG_ONE: Self = Self {
+        mantissa: IBig::NEG_ONE,
+        exponent: 0,
+        precision: 1,
+    };
+    /// [FloatRepr] representing positive infinity (∞)
+    pub const INFINITY: Self = Self {
+        mantissa: IBig::ZERO,
+        exponent: 1,
+        precision: 0,
+    };
+    /// [FloatRepr] representing negative infinity (-∞)
+    pub const NEG_INFINITY: Self = Self {
+        mantissa: IBig::ZERO,
+        exponent: -1,
+        precision: 0,
+    };
+
+    pub const fn is_zero(&self) -> bool {
+        self.mantissa.is_zero() && self.exponent == 0
     }
 
-    #[inline]
-    pub const fn one() -> Self {
-        Self { mantissa: IBig::one(), exponent: 0, precision: 1 }
+    pub const fn is_one(&self) -> bool {
+        self.mantissa.is_one() && self.exponent == 0
     }
 
-    #[inline]
-    pub const fn neg_one() -> Self {
-        Self { mantissa: IBig::neg_one(), exponent: 0, precision: 1 }
+    pub const fn is_infinite(&self) -> bool {
+        self.mantissa.is_zero() && self.exponent != 0
+    }
+
+    pub const fn is_finite(&self) -> bool {
+        !self.is_infinite()
     }
 
     fn ceil(&self) -> IBig {