+++ integer/src/modular/pow.rs
@@ -20,11 +20,11 @@ impl<'a> Modulo<'a> {
     /// ```
     /// # use dashu_int::{modular::ModuloRing, UBig};
     /// // A Mersenne prime.
-    /// let p = UBig::from(2u8).pow(607) - UBig::one();
+    /// let p = UBig::from(2u8).pow(607) - UBig::ONE;
     /// let ring = ModuloRing::new(p.clone());
     /// // Fermat's little theorem: a^(p-1) = 1 (mod p)
     /// let a = ring.convert(123);
-    /// assert_eq!(a.pow(&(p - UBig::one())), ring.convert(1));
+    /// assert_eq!(a.pow(&(p - UBig::ONE)), ring.convert(1));
     /// ```
     #[inline]
     pub fn pow(&self, exp: &UBig) -> Modulo<'a> {