+++ macros/tests/int.rs
@@ -6,8 +6,8 @@ use dashu_macros::{ibig, ubig};
 #[test]
 fn test_ubig() {
     // decimals
-    assert_eq!(ubig!(0), UBig::zero());
-    assert_eq!(ubig!(00000001), UBig::one());
+    assert_eq!(ubig!(0), UBig::ZERO);
+    assert_eq!(ubig!(00000001), UBig::ONE);
     assert_eq!(ubig!(12341234), UBig::from(12341234u32));
     assert_eq!(ubig!(12_34_12_34_), UBig::from(12_34_12_34_u32));
     assert_eq!(
@@ -20,9 +20,9 @@ fn test_ubig() {
     );
 
     // hexadecimals
-    assert_eq!(ubig!(0 base 16), UBig::zero());
-    assert_eq!(ubig!(1 base 16), UBig::one());
-    assert_eq!(ubig!(0x00000001), UBig::one());
+    assert_eq!(ubig!(0 base 16), UBig::ZERO);
+    assert_eq!(ubig!(1 base 16), UBig::ONE);
+    assert_eq!(ubig!(0x00000001), UBig::ONE);
     assert_eq!(ubig!(0x12341234), UBig::from(0x12341234u32));
     assert_eq!(ubig!(abcdef base 16), UBig::from(0xabcdefu32));
     assert_eq!(
@@ -62,12 +62,12 @@ fn test_ubig() {
 #[test]
 fn test_ibig() {
     // decimals
-    assert_eq!(ibig!(0), IBig::zero());
-    assert_eq!(ibig!(+0), IBig::zero());
-    assert_eq!(ibig!(-0), IBig::zero());
-    assert_eq!(ibig!(00000001), IBig::one());
-    assert_eq!(ibig!(+00000001), IBig::one());
-    assert_eq!(ibig!(-00000001), IBig::neg_one());
+    assert_eq!(ibig!(0), IBig::ZERO);
+    assert_eq!(ibig!(+0), IBig::ZERO);
+    assert_eq!(ibig!(-0), IBig::ZERO);
+    assert_eq!(ibig!(00000001), IBig::ONE);
+    assert_eq!(ibig!(+00000001), IBig::ONE);
+    assert_eq!(ibig!(-00000001), IBig::NEG_ONE);
     assert_eq!(ibig!(-12341234), IBig::from(-12341234));
     assert_eq!(ibig!(-12_34_12_34_), IBig::from(-12_34_12_34_));
     assert_eq!(
@@ -88,15 +88,15 @@ fn test_ibig() {
     );
 
     // hexadecimals
-    assert_eq!(ibig!(0 base 16), IBig::zero());
-    assert_eq!(ibig!(+0 base 16), IBig::zero());
-    assert_eq!(ibig!(-0 base 16), IBig::zero());
-    assert_eq!(ibig!(1 base 16), IBig::one());
-    assert_eq!(ibig!(+1 base 16), IBig::one());
-    assert_eq!(ibig!(-1 base 16), IBig::neg_one());
-    assert_eq!(ibig!(0x00000001), IBig::one());
-    assert_eq!(ibig!(+0x00000001), IBig::one());
-    assert_eq!(ibig!(-0x00000001), IBig::neg_one());
+    assert_eq!(ibig!(0 base 16), IBig::ZERO);
+    assert_eq!(ibig!(+0 base 16), IBig::ZERO);
+    assert_eq!(ibig!(-0 base 16), IBig::ZERO);
+    assert_eq!(ibig!(1 base 16), IBig::ONE);
+    assert_eq!(ibig!(+1 base 16), IBig::ONE);
+    assert_eq!(ibig!(-1 base 16), IBig::NEG_ONE);
+    assert_eq!(ibig!(0x00000001), IBig::ONE);
+    assert_eq!(ibig!(+0x00000001), IBig::ONE);
+    assert_eq!(ibig!(-0x00000001), IBig::NEG_ONE);
     assert_eq!(ibig!(-0x12341234), IBig::from(-0x12341234));
     assert_eq!(ibig!(-abcdef base 16), IBig::from(-0xabcdef));
     assert_eq!(